added legend to two bar chart
