+++ Swift Charts Examples/Charts/BarCharts/Two Bars/TwoBars.swift
@@ -1,5 +1,5 @@
 //
-//  TwoBarsDetail.swift
+//  TwoBarsSimple.swift
 //  Swift Charts Examples
 //
 //  Created by Jordi Bruin on 12/06/2022.
@@ -8,13 +8,48 @@
 import SwiftUI
 import Charts
 
+struct TwoBarsSimpleOverview: View {
+    var body: some View {
+        VStack(alignment: .leading) {
+            Text("Two Bars")
+                .font(.callout)
+                .foregroundStyle(.secondary)
 
+            Chart {
+                ForEach(LocationData.last30Days) { series in
+                    ForEach(series.sales, id: \.weekday) { element in
+                        BarMark(
+                            x: .value("Day", element.weekday, unit: .day),
+                            y: .value("Sales", element.sales)
+                        )
+                        .accessibilityLabel("\(element.weekday.formatted())")
+                        .accessibilityValue("\(element.sales)")
+                    }
+                    .foregroundStyle(by: .value("City", series.city))
+//                    .symbol(by: .value("City", series.city))
+                }
+            }
+            .chartXAxis(.hidden)
+            .chartYAxis(.hidden)
+            .chartLegend(.hidden)
+                .frame(height: 100)
+        }
+    }
+}
+
+struct TwoBarsOverview_Previews: PreviewProvider {
+    static var previews: some View {
+        TwoBarsSimpleOverview()
+            .padding()
+    }
+}
 
 struct TwoBarsSimpleDetailView: View {
 
     @State var lineWidth = 2.0
     @State var interpolationMethod: ChartInterpolationMethod = .cardinal
     @State var strideBy: ChartStrideBy = .day
+    @State var showLegend = false
 
     var body: some View {
         List {
@@ -39,6 +74,7 @@ struct TwoBarsSimpleDetailView: View {
                         AxisValueLabel(format: .dateTime.weekday(.abbreviated), centered: true)
                     }
                 }
+                .chartLegend(showLegend ? .visible : .hidden)
                 .chartLegend(position: .top)
                 .frame(height: 240)
             }
@@ -60,14 +96,7 @@ struct TwoBarsSimpleDetailView: View {
                 }
             }
             
-            Picker("Interpolation Method", selection: $interpolationMethod) {
-                ForEach(ChartInterpolationMethod.allCases) { Text($0.mode.description).tag($0) }
-            }
-            
-//            Picker("Stride By", selection: $strideBy) {
-//                ForEach(ChartStrideBy.allCases) { Text($0.title).tag($0)
-//                }
-//            }
+            Toggle("Show Chart Legend", isOn: $showLegend)
         }
     }
 }
